- fixed some warnings with SIZE_T conversions
- renamed scylla_wrapper header file

+++ TitanEngine/stdafx.h
@@ -114,7 +114,7 @@ typedef struct
 {
     BYTE BreakPointActive;
     ULONG_PTR BreakPointAddress;
-    DWORD BreakPointSize;
+    SIZE_T BreakPointSize;
     BYTE OriginalByte[10];
     int BreakPointType;
     int AdvancedBreakPointType;
